⬆️ Update dependencies and package manager
Upgrade next from 15.0.2 to 15.0.3. Update bun packageManager
from 1.1.33 to 1.1.34. Upgrade @aws-sdk packages from 3.685.0
to 3.686.0. Update @neondatabase/serverless from 0.10.2 to 0.10.3.
Upgrade drizzle-orm from 0.35.0 to 0.36.1 and drizzle-kit
from 0.26.2 to 0.28.0.

These updates aim to incorporate the latest features,
improvements, and bug fixes, ensuring that the project
remains up-to-date and secure with the recent releases.